backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg[binary]>=3.2.0
pgvector>=0.2.4
pydantic>=2.5.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
umap-learn>=0.5.5
numpy>=1.26.0
pandas>=2.1.0 